+++ src/main/java/com/example/backendsynergyproject/dto/VersionDto.java
@@ -1,6 +1,6 @@
 package com.example.backendsynergyproject.dto;
 
 public record VersionDto(
-        Integer percentOfSpendBudget,
+        Float percentOfSpendBudget,
         String name
 ) {}
